Change dp_ram design to match PuLPino more closely
This switches back most of the changes to the dp_ram that were made
since incorporating it into the verilator model. The dp_ram is now
similar to the one in upstream PuLPino, with the exception that it
divides addresses by 4 before accessing memory because the core provides
byte-addresses whilst the memory uses word addresses. According to the
LSU documentation, this is not an issue because non-word-aligned
accesses are split into two word-aligned accesses the the appropriate
byte-enable bits set.

Files changed:

        * dp_ram.sv: Declare mem as [3:0][7:0] again, divide addresses
        by 4 for accessing it. (readByte, writeByte) Perform required
        division and modulo operations to access individual bytes of
        mem.